Move what was left of bridge.js to matrix-to-irc.js
