+++ lib/irc-appservice.js
@@ -7,7 +7,7 @@ module.exports.serviceName = "matrix-appservice-irc";
 
 var q = require("q");
 
-var bridge = require("./bridge.js");
+var matrixToIrc = require("./matrix-to-irc.js");
 var ircToMatrix = require("./irc-to-matrix.js");
 var IrcServer = require("./irclib/server.js").IrcServer;
 var ircLib = require("./irclib/irc.js");
@@ -71,8 +71,8 @@ module.exports.register = function(controller, serviceConfig) {
     controller.setLogger(function(line) {
         log.info(line.replace(/\n/g, " "));
     });
-    controller.setAliasQueryResolver(bridge.hooks.matrix.onAliasQuery);
-    controller.setUserQueryResolver(bridge.hooks.matrix.onUserQuery);
+    controller.setAliasQueryResolver(matrixToIrc.onAliasQuery);
+    controller.setUserQueryResolver(matrixToIrc.onUserQuery);
     for (var i = 0; i < servers.length; i++) {
         var server = servers[i];
         // add an alias pattern for servers who want aliases exposed.
@@ -160,8 +160,8 @@ module.exports.register = function(controller, serviceConfig) {
         defer.reject(err);
     });
 
-    controller.on("type:m.room.message", bridge.hooks.matrix.onMessage);
-    controller.on("type:m.room.topic", bridge.hooks.matrix.onMessage);
+    controller.on("type:m.room.message", matrixToIrc.onMessage);
+    controller.on("type:m.room.topic", matrixToIrc.onMessage);
     controller.on("type:m.room.member", function(event) {
         if (!event.content || !event.content.membership) {
             return;
@@ -169,13 +169,13 @@ module.exports.register = function(controller, serviceConfig) {
         var target = new MatrixUser(event.state_key, null, null);
         var sender = new MatrixUser(event.user_id, null, null);
         if (event.content.membership === "invite") {
-            return bridge.hooks.matrix.onInvite(event, sender, target);
+            return matrixToIrc.onInvite(event, sender, target);
         }
         else if (event.content.membership === "join") {
-            return bridge.hooks.matrix.onJoin(event, target);
+            return matrixToIrc.onJoin(event, target);
         }
         else if (["ban", "leave"].indexOf(event.content.membership) !== -1) {
-            return bridge.hooks.matrix.onLeave(event, target);
+            return matrixToIrc.onLeave(event, target);
         }
     });
