+++ lib/membershiplists.js
@@ -3,7 +3,7 @@
 "use strict";
 
 var q = require("q");
-var bridge = require("./bridge.js");
+var matrixToIrc = require("./matrix-to-irc.js");
 var ircLib = require("./irclib/irc.js");
 var matrixLib = require("./mxlib/matrix");
 var MatrixUser = require("./models/users").MatrixUser;
@@ -240,7 +240,7 @@ function getRoomMemberData(server, roomId, stateEvents) {
 
 function injectJoinEvent(roomId, userId) {
     var target = new MatrixUser(userId, null, null);
-    return bridge.hooks.matrix.onJoin({
+    return matrixToIrc.onJoin({
         event_id: "$fake:membershiplist",
         room_id: roomId,
         state_key: userId,
